very hacky dev mode to do a dry run of build

+++ script/build.js
@@ -60,5 +60,10 @@ let allFilesCategorized = {},
     root.categories = allFilesCategorized;
 
     console.log(JSON.stringify(root, null, 2));
-    fs.writeFileSync('./api/docs.json', JSON.stringify(root, null, 4));
+    const dryRun = ['-n', '--dry-run'].map(v => process.argv.includes(v)).includes(true);
+    if (!dryRun) {
+        fs.writeFileSync('./api/docs.json', JSON.stringify(root, null, 4));
+    } else {
+        console.info('Dry Run: skipping writing output to ./api/docs.json.');
+    }
 })();
